+++ tests/integration/cases/call_wrap_in_file/Bash_call.sh
@@ -1,0 +1,3 @@
+process() { :; }
+process 1 2
+process 3 4