+++ tests/integration/cases/call_wrap_in_file/Erlang_call.erl
@@ -1,0 +1,6 @@
+-module(check).
+-export([x/0]).
+process(_, _) -> ok.
+x() ->
+    process(1, 2),
+    process(3, 4).