+++ tests/integration/cases/call_wrap_in_file/Kotlin_call.kts
@@ -1,0 +1,3 @@
+fun process(a: Any? = null, b: Any? = null): Any? = null
+process(a = 1, b = 2)
+process(a = 3, b = 4)