+++ tests/integration/cases/call_wrap_in_file/Php_call.php
@@ -1,0 +1,4 @@
+<?php
+function process($a, $b) {}
+process(a: 1, b: 2);
+process(a: 3, b: 4);