+++ tests/integration/cases/call_wrap_in_file/Racket_call.rkt
@@ -1,0 +1,4 @@
+#lang racket
+(define process (make-keyword-procedure (lambda _ (void))))
+(process #:a 1 #:b 2)
+(process #:a 3 #:b 4)